requirements.txt
futu-api==9.4.5408
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.9.2
seaborn>=0.13.2
psutil>=7.1.2  # 新增
pyyaml>=6.0.1
pytz>=2024.1
ccxt>=4.5.14
python-binance==1.0.32